requirements.txt
earthengine_api==0.1.397
google_api_python_client==2.115.0
google_auth_oauthlib==1.0.0
numpy==1.26.4
protobuf==5.26.1
